+++ src/pages/species-richness/components/species-richness-table/species-richness-table.ts
@@ -3,26 +3,42 @@ import { Prop } from 'vue-property-decorator'
 
 import { ChartModels } from '@/models'
 import ExportButtonView from '@/views/export-button.vue'
+import NoDataContainerView from '@/views/no-data-container.vue'
+
+interface Header {
+  title: string
+  color: string
+}
+
+const HEADER_COLOR = '#ffffff80'
 
 @Options({
   components: {
-    ExportButtonView
+    ExportButtonView,
+    NoDataContainerView
   }
 })
 export default class SpeciesRichnessTable extends Vue {
   @Prop({ default: [] }) tableData!: ChartModels.TableData[]
+  @Prop({ default: [] }) colors!: string[]
 
-  get tableHeader (): string[] {
+  get tableHeader (): Header[] {
     return [
-      'Species',
-      'Class',
-      ...Array.from({ length: this.datasetCount }, (v, i) => `Dataset ${i + 1}`),
-      'Total'
+      { title: 'Species', color: HEADER_COLOR },
+      { title: 'Class', color: HEADER_COLOR },
+      ...((this.hasMoreThanOneDataset)
+        ? [...Array.from({ length: this.datasetCount }, (v, i) => ({ title: `Dataset ${i + 1}`, color: this.colors[i] })), { title: 'Total', color: HEADER_COLOR }]
+        : []
+      )
     ]
   }
 
   get hasTableData (): boolean {
     return this.tableData.length > 0
+  }
+
+  get hasMoreThanOneDataset (): boolean {
+    return this.datasetCount > 1
   }
 
   get datasetCount (): number {