+++ src/pages/species-richness/species-richness.ts
@@ -29,6 +29,7 @@ dayjs.extend(utc)
 export default class SpeciesRichnessPage extends Vue {
   sites: SiteModels.Site[] = []
 
+  colors: string[] = []
   filters: SpeciesRichnessFilter[] = []
   detectionCounts: number[] = []
   chartData: ChartModels.GroupedBarChartItem[] = []
@@ -52,6 +53,7 @@ export default class SpeciesRichnessPage extends Vue {
     )
 
     this.filters = filters
+    this.colors = datasets.map(ds => ds.color)
     this.detectionCounts = datasets.map(ds => ds.data.detectionCount)
     this.chartData = this.getBarChartDataset(datasets)
     this.mapDatasets = this.getMapDataset(datasets)
@@ -97,14 +99,10 @@ export default class SpeciesRichnessPage extends Vue {
     }))
 
     // TODO - 106: Update filename and folder name
-    const files: FileModels.File[] = []
-    for (const [idx, csv] of csvs.entries()) {
-      const base64 = await FileUtils.generateBase64Sheet(csv, 'csv', 'Species Report')
-      files.push({
-        filename: `report-${idx + 1}.csv`,
-        data: base64
-      })
-    }
+    const files: FileModels.File[] = await Promise.all(csvs.map(async (csv, idx) => ({
+      filename: `report-${idx + 1}.csv`,
+      data: await FileUtils.generateSheet(csv, 'csv', 'Species Report')
+    })))
     await FileUtils.zipFiles(files, 'reports')
   }
 }