+++ CHANGELOG.md
@@ -23,3 +23,5 @@
 * **features:** Export species richness raw data as csv (#36)
 * **features:** Activity patterns species selector (#89)
 * **features:** Export csv support multiple datasets(#106)
+
+* **refactor:** Refactor species richness tabledisplay (#112)